Unify file exceptions
