--- src/Exception/File/MutationExceptionInterface.php
@@ -1,12 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Exception\File;
-
-use League\Flysystem\FilesystemException;
-
-interface MutationExceptionInterface extends FileExceptionInterface
-{
-    public function getFilesystemException(): FilesystemException;
-}
