--- src/Exception/File/PathExceptionInterface.php
@@ -1,10 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Exception\File;
-
-interface PathExceptionInterface extends FileExceptionInterface
-{
-    public function getPath(): string;
-}
