+++ src/Exception/File/ReadException.php
@@ -6,26 +6,10 @@ namespace App\Exception\File;
 
 use League\Flysystem\FilesystemException;
 
-class ReadException extends \Exception implements PathExceptionInterface
+class ReadException extends AbstractFileException
 {
-    public function __construct(
-        private string $path,
-        private FilesystemException $filesystemException,
-    ) {
-        parent::__construct(
-            sprintf('Unable to read file "%s"', $path),
-            0,
-            $this->filesystemException
-        );
-    }
-
-    public function getPath(): string
+    public function __construct(string $path, FilesystemException $filesystemException)
     {
-        return $this->path;
-    }
-
-    public function getFilesystemException(): FilesystemException
-    {
-        return $this->filesystemException;
+        parent::__construct('read', $path, $filesystemException);
     }
 }
