+++ src/Exception/File/RemoveException.php
@@ -6,26 +6,10 @@ namespace App\Exception\File;
 
 use League\Flysystem\FilesystemException;
 
-class RemoveException extends \Exception implements MutationExceptionInterface, PathExceptionInterface
+class RemoveException extends AbstractFileException
 {
-    public function __construct(
-        private string $path,
-        private FilesystemException $filesystemException,
-    ) {
-        parent::__construct(
-            sprintf('Unable to remove "%s"', $path),
-            0,
-            $this->filesystemException
-        );
-    }
-
-    public function getPath(): string
+    public function __construct(string $path, FilesystemException $filesystemException)
     {
-        return $this->path;
-    }
-
-    public function getFilesystemException(): FilesystemException
-    {
-        return $this->filesystemException;
+        parent::__construct('remove', $path, $filesystemException);
     }
 }
